Update the method reference to handle method calls with parameters
- This fix an issue with the reference logic preventing the user
for jumping to bean method when using parameters in the bean call.

The reference logic is limit to show all the matching method names regardless
it's a parameter references. Later we add more support to link it more
accurately to the method with parameters it should match.

- Fix an issue with ambiguous methods was mark, if the bean method call
was with parameters. Currently we are not able to accurately match the
bean method call. This will come later.