Ref #908: Camel Debugger - Add support of Camel SpringBoot 4 (#913)
## Motivation

Several problems prevent the Camel Debugger from working properly with Camel SpringBoot 4 applications. 

If the debugger is launched using a Java run configuration, then the Camel Debugger will be able to connect to the Camel SpringBoot application using JMX, but an error indicating that `camel-spring-xml` must added to the classpath is thrown. 

If the debugger is launched using the Camel run configuration for SpringBoot, then the Camel Debugger won't be able to connect as the Camel SpringBoot application is launched in a different process. 

## Modifications:

* Allow to add an additional artifact when configuring the debugger to also add `camel-spring-xml` for SpringBoot
* Access to JMX using a JMX service URL in case of SpringBoot
* Avoid Gradle issues when the generated parameters have not been removed from the Gradle configuration
* Avoid NPE when a Camel checkpoint has been added to a Java class and the Java class is then commented out